use common configuration
